Move OTU FASTA download endpoint ot '/api/otus/:id.fa'
