+++ virtool/downloads/db.py
@@ -3,7 +3,7 @@ Functions for working with the database to provide file downloads. All functions
 child models.
 
 """
-from typing import Tuple
+from typing import Tuple, Optional
 
 import virtool.downloads.utils
 import virtool.errors
@@ -64,7 +64,7 @@ async def generate_sequence_fasta(db, sequence_id: str) -> Tuple[str, str]:
     return virtool.downloads.utils.format_fasta_filename(otu_name, isolate_name, sequence["_id"]), fasta
 
 
-async def generate_otu_fasta(db, otu_id: str) -> Tuple[str, str]:
+async def generate_otu_fasta(db, otu_id: str) -> Tuple[Optional[str], Optional[str]]:
     """
     Generate a FASTA filename and body for the sequences associated with the otu identified by the passed
     ``otu_id``.
@@ -76,19 +76,20 @@ async def generate_otu_fasta(db, otu_id: str) -> Tuple[str, str]:
     """
     otu = await db.otus.find_one(otu_id, ["name", "isolates"])
 
-    if not otu:
-        raise virtool.errors.DatabaseError("OTU does not exist")
-
     fasta = list()
 
     for isolate in otu["isolates"]:
-        async for sequence in db.sequences.find({"otu_id": otu_id, "isolate_id": isolate["id"]}, ["sequence"]):
-            fasta.append(virtool.downloads.utils.format_fasta_entry(
-                otu["name"],
-                virtool.otus.utils.format_isolate_name(isolate),
-                sequence["_id"],
-                sequence["sequence"]
-            ))
+        sequences = await db.sequences.find({"otu_id": otu_id, "isolate_id": isolate["id"]}, ["sequence"]).to_list(None)
+        if sequences:
+            for sequence in sequences:
+                fasta.append(virtool.downloads.utils.format_fasta_entry(
+                    otu["name"],
+                    virtool.otus.utils.format_isolate_name(isolate),
+                    sequence["_id"],
+                    sequence["sequence"]
+                ))
+        else:
+            return None, None
 
     fasta = "\n".join(fasta)
